Revise Voxel functionalities in VoxelHashMap and Preprocessing (#362)
* Unify Voxel functionalities, now everything comes from VoxelHashMap

* This looks like an even better unification

* Modernize VoxelDownsample

* VoxelUtils is born, now everthing is liteally in one place, and we have
some side effect code reduction

* Macos does not like my include

* Local style include

* With move

* As of benchmark

* For each for the win

* Remove unnecessary include

* Const correcteness

* Refresh changes

---------
